requirements.txt
appdirs==1.4.1
click==6.7
colorama==0.3.7
coreapi==2.2.4
coreapi-cli==1.0.6
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.10.5
djangorestframework==3.5.4
gunicorn==19.6.0
httpie==0.9.9
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
psycopg2==2.6.2
Pygments==2.2.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
whitenoise==3.3.0
